requirements.txt
numpy
pandas
feltlabs==0.5.1 
PyNaCl==1.5.0
matplotlib
seaborn

# Dev
jupyterlab
pre-commit
black